fix: replace Novela with RobotoSerif
Novela is a broken font, which has an incorrect font feature table
which makes ligatures unusable on the web. Politely asking the typeface
designers to let me fix it, led to a denial on their side and refusal
to let me modify the version to correct it. Cannot recommend that
foundry.